* as/link/lkar.h: changed type of ar_size to size_t to make it work on   ppc Mac OS X (off_t is 64 bit and is not converted to 32 bit   unsigned int when calling new(), since asxxxx doesn't use function   prototypes)
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@5347 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ as/link/lkar.h
@@ -75,7 +75,7 @@ struct ar_hdr                     /* archive member header */
   uid_t ar_uid;                   /* archive member user identification */
   gid_t ar_gid;                   /* archive member group identification */
   mode_t ar_mode;                 /* archive member mode (octal) */
-  off_t ar_size;                  /* archive member size */
+  size_t ar_size;                 /* archive member size */
 };
 
 #endif /* __LKAR_H */
